// apps/web/src/api/generated/WorkflowVersionDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type WorkflowVersionDto = { workflow_id: string, version: string, canonical_hash: string, author_kind: string, extension_id: string | null, extension_version: string | null, created_at: string, is_current: boolean, };